Moving out DocumentIDStrategy as a top level enum
